requirements.txt
torch
torchvision
numpy
pillow
tensorboardX
opencv-python
plyfile